requirements.txt
python-dotenv>=1.0.0
gdown>=5.0.0
flask>=3.0.0
cryptography>=41.0.0
PyMySQL>=1.1.0
gunicorn>=21.0.0
